fix: inline checkRentalOwnershipForQr to fix Next.js server export error
Next.js 'use server' files don't allow re-exporting functions.
Moved the function inline into rentals.ts.
